Update demo to use latest
