+++ app/javascript/controllers/todos_controller.js
@@ -1,8 +1,37 @@
-import { Controller } from 'stimulus';
+import ApplicationController from './application_controller';
+import Velocity from 'velocity-animate';
 
-export default class extends Controller {
+export default class extends ApplicationController {
+  static targets = ['list'];
+
   cancelEdit(event) {
     if (event.type === 'keydown' && !['Escape', 'Esc'].includes(event.key)) return;
     this.stimulate('TodosReflex#cancel_edit');
+  }
+
+  afterCreate() {
+    this.highlight(this.lastListItem);
+  }
+
+  afterToggle(checkbox) {
+    this.highlight(checkbox.closest('li'));
+  }
+
+  afterToggleAll() {
+    this.listItems.forEach(li => this.highlight(li));
+  }
+
+  highlight(element) {
+    Velocity(element, { scale: 1.01, backgroundColor: '#ffff99' }, 200).then(
+      Velocity(element, { scale: 1, backgroundColor: '#fff' }, 600)
+    );
+  }
+
+  get listItems() {
+    return this.listTarget.querySelectorAll('li');
+  }
+
+  get lastListItem() {
+    return this.listItems[this.listItems.length - 1];
   }
 }
