+++ app/views/todos/_todo.html.erb
@@ -1,7 +1,6 @@
 <% if todo.id == @edit_id %>
   <li class="editing">
     <input autofocus type="text" class="edit" value="<%= todo.title %>"
-      data-controller="todos"
       data-action="keydown->todo#cancelEdit blur->todo#cancelEdit"
       data-reflex="change->TodosReflex#update" data-id="<%= todo.id %>" data-room="<%= session.id %>">
   </li>
