+++ app/views/todos/index.html.erb
@@ -1,4 +1,4 @@
-<section class="todoapp" data-controller="visual-effects">
+<section class="todoapp" data-controller="todos">
   <header class="header">
     <h1><small><%= truncate "#{session[:filter]} todos", length: 12 %></small></h1>
     <input <%= "autofocus " unless @edit_id %> class="new-todo" placeholder="What needs to be done?"
